backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
yfinance>=1.1.0
pandas>=2.3.2
pandas-ta==0.4.71b0
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.20
apscheduler==3.10.4
